+++ src/salmon/validation.clj
@@ -25,20 +25,29 @@
 (defn- refs [x]
   (sp/select (sp/walker ds/ref?) x))
 
-(defn- resolve-ref [system ref]
+(defn- resolve-ref [system referencing-component-id ref]
   (let [{::ds/keys [instances resolved-defs]} system
-        k (ds/ref-key ref)
-        k2 (take 2 k)
-        instance (get-in instances k2)
-        def (get-in resolved-defs k2)]
+        rt (ds/ref-type ref)
+        rkey (cond
+               (= ::ds/ref rt)
+               (take 2 (ds/ref-key ref))
+
+               (= ::ds/local-ref rt)
+               (into [(first referencing-component-id)]
+                     (ds/ref-key ref))
+
+               :else
+               (throw (ex-info "Not a ref" {:value ref})))
+        instance (get-in instances rkey)
+        def (get-in resolved-defs rkey)]
     (cond
-      instance (get-in instance (drop 2 k))
-      def (::ds/start (get-in def (drop 2 k))))))
+      instance (get-in instance (drop 2 rkey))
+      def (::ds/start (get-in def (drop 2 rkey))))))
 
 (defn resolve-refs
   "Resolve all refs in x."
-  [system x]
-  (sp/transform (sp/walker ds/ref?) (partial resolve-ref system) x))
+  [system referencing-component-id x]
+  (sp/transform (sp/walker ds/ref?) (partial resolve-ref system referencing-component-id) x))
 
 (defn- ref-resolveable? [system component-id ref]
   (let [{::ds/keys [instances resolved-defs]} system